+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/PodIT.java
@@ -46,6 +46,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
+import java.util.Locale;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -78,7 +79,7 @@ public class PodIT {
     client.pods().inNamespace(currentNamespace).delete();
 
     pod1 = new PodBuilder()
-      .withNewMetadata().withName("pod1-" + RandomStringUtils.randomAlphanumeric(6).toLowerCase()).endMetadata()
+      .withNewMetadata().withName("pod1-" + RandomStringUtils.randomAlphanumeric(6).toLowerCase(Locale.ROOT)).endMetadata()
       .withNewSpec()
       .addNewContainer()
       .withName("busybox")