+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Fix #1827: Fix `withGracePeriod` and `withPropagationPolicy` return type to safely chain further DSL methods and default GracePeriod to 30s
 * Fix #1828: VersionInfo date parsing of year
 * Fix #1838: Use the correct apiGroup for Knative in KnativeResourceMappingProvider
-
+* Fix #1844: KubernetesDeserializer can now handle ArrayNode.
 
 #### Improvements
 * Cleanup log4j dependency from project